refactor to amadaues
